+++ CHANGELOG.md
@@ -900,6 +900,10 @@ Implementing this changes involved breaking some existing functionality:
 
 * `qc/calculate_qc_metrics`: fix calculating mitochondrial gene related QC metrics when only or no mitochondrial genes were found (PR #564).
 
+## NEW FUNCTIONALITY
+
+* Added `protein_processing/dsb_index` and `protein_processing/dsb_normalize` components (PR #588).
+
 # openpipelines 0.10.1
 
 ## MINOR CHANGES